+++ src/common/toasts/EmailToast.css
