+++ src/common/toasts/EmailToast.js
